+++ contracts/interface/IProduct.sol
@@ -18,8 +18,8 @@ interface IProduct {
     function governance() external view returns (address);
     function newGovernance() external view returns (address);
     function price() external view returns (uint24);
-    function minPeriod() external view returns (uint64);
-    function maxPeriod() external view returns (uint64);
+    function minPeriod() external view returns (uint40);
+    function maxPeriod() external view returns (uint40);
     function maxCoverAmount() external view returns (uint256);
     function maxCoverPerUser() external view returns (uint256);
     function maxCoverPerUserDivisor() external view returns (uint32);
@@ -30,8 +30,8 @@ interface IProduct {
     function setGovernance(address _governance) external;
     function acceptGovernance() external;
     function setPrice(uint24 _price) external;
-    function setMinPeriod(uint64 _minPeriod) external;
-    function setMaxPeriod(uint64 _maxPeriod) external;
+    function setMinPeriod(uint40 _minPeriod) external;
+    function setMaxPeriod(uint40 _maxPeriod) external;
     //function setMaxCoverPerUserDivisor(uint32 _maxCoverPerUserDivisor) external;
     function setCoveredPlatform(address _coveredPlatform) external;
     function setPolicyManager(address _policyManager) external;
@@ -45,16 +45,16 @@ interface IProduct {
     /**** QUOTE VIEW FUNCTIONS
     View functions that give us quotes regarding a policy
     ****/
-    function getQuote(address _policyholder, address _positionContract, uint256 _coverAmount, uint64 _blocks) external view returns (uint256);
+    function getQuote(address _policyholder, address _positionContract, uint256 _coverAmount, uint40 _blocks) external view returns (uint256);
 
     /**** MUTATIVE FUNCTIONS
     Functions that deploy and change policy contracts
     ****/
     function updateActiveCoverAmount(int256 _coverDiff) external;
-    function buyPolicy(address _policyholder, address _positionContract, uint256 _coverAmount, uint64 _blocks) external payable returns (uint256 policyID);
+    function buyPolicy(address _policyholder, address _positionContract, uint256 _coverAmount, uint40 _blocks) external payable returns (uint256 policyID);
     function updateCoverAmount(uint256 _policyID, uint256 _coverAmount) external payable;
-    function extendPolicy(uint256 _policyID, uint64 _blocks) external payable;
+    function extendPolicy(uint256 _policyID, uint40 _blocks) external payable;
     function cancelPolicy(uint256 _policyID) external;
-    function updatePolicy(uint256 _policyID, uint256 _coverAmount, uint64 _blocks ) external payable;
+    function updatePolicy(uint256 _policyID, uint256 _coverAmount, uint40 _blocks ) external payable;
 
 }
