+++ contracts/products/CompoundProduct.sol
@@ -30,8 +30,8 @@ contract CompoundProduct is BaseProduct, EIP712 {
         IPolicyManager _policyManager,
         IRegistry _registry,
         address _coveredPlatform,
-        uint64 _minPeriod,
-        uint64 _maxPeriod,
+        uint40 _minPeriod,
+        uint40 _maxPeriod,
         uint24 _price,
         uint32 _maxCoverPerUserDivisor,
         address _quoter
