+++ contracts/products/CurveProduct.sol
@@ -31,8 +31,8 @@ contract CurveProduct is BaseProduct {
         IPolicyManager _policyManager,
         IRegistry _registry,
         address _coveredPlatform,
-        uint64 _minPeriod,
-        uint64 _maxPeriod,
+        uint40 _minPeriod,
+        uint40 _maxPeriod,
         uint24 _price,
         uint32 _maxCoverPerUserDivisor,
         address _quoter
@@ -96,5 +96,5 @@ contract CurveProduct is BaseProduct {
     function name() public pure override returns (string memory) {
         return "Curve";
     }
-    
+
 }
